refactor(language): restructure error handling and validation
Refactor error handling by introducing `UnimplementedError` and `MultipleError` classes to replace the previous `SingleError` and `UnimplementedSingleError`. Update validation logic to use the new error classes and rename `ExceptionCollector` to improve clarity. This change enhances maintainability and simplifies the error handling structure.
