+++ src/index.ts
@@ -3,4 +3,4 @@ export * from "./GraplixIdentifier";
 export * from "./GraplixInput";
 export * from "./GraplixResolvers";
 export * from "./GraplixSchema";
-export * from "./language";
+export { fga } from "./language";
