+++ src/language/UnimplementedError.ts
@@ -1,0 +1,11 @@
+enum ErrorCode {
+  UNIMPLEMENTED = "UNIMPLEMENTED",
+}
+
+export class UnimplementedError extends Error {
+  code = ErrorCode.UNIMPLEMENTED;
+
+  constructor(featureName: string) {
+    super(`${featureName} is not implemented`);
+  }
+}
