+++ src/language/index.ts
@@ -1,2 +1,1 @@
 export * from "./fga";
-export * from "./errors";
