+++ src/utils/ExceptionCollector.ts
@@ -1,0 +1,9 @@
+import { UnimplementedError } from "../language/UnimplementedError";
+
+export class ExceptionCollector {
+  constructor(private errors: UnimplementedError[]) {}
+
+  captureUnimplementedError(featureName: string) {
+    this.errors.push(new UnimplementedError(featureName));
+  }
+}
