--- src/utils/exceptions.ts
@@ -1,9 +1,0 @@
-import { UnimplementedSingleError } from "../language";
-
-export class ExceptionCollector {
-  constructor(private errors: UnimplementedSingleError[]) {}
-
-  captureUnimplementedError(featureName: string) {
-    this.errors.push(new UnimplementedSingleError(featureName));
-  }
-}
